requirements.txt
# BERU 2.0 Dependencies

# Core
pyyaml>=6.0
aiohttp>=3.9.0

# Memory (optional - for vector storage)
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.0.0
